+++ src/Flex/Filter/SanitizeFilename.php
@@ -13,12 +13,15 @@ class SanitizeFilename {
      * @return string
      */
     public function filter($filename) {
-        // replace path
-        $filename = str_replace('..', '', $filename);
-        $filename = str_replace('/', '', $filename);
+        $extension = pathinfo($filename, PATHINFO_EXTENSION);
+
+        // remove possible suffix from path
+        if(!empty($extension) && !preg_match('/\s/', $extension)) {
+            $filename = substr($filename, 0, strrpos($filename, "."));
+        }
 
         // replace all invalid chars with whitespaces
-        $filename = preg_replace('/[^a-zA-Z0-9-_]/', ' ', $filename);
+        $filename = preg_replace('/[^a-zA-Z0-9]/', ' ', $filename);
 
         // replace multiple whitespaces to single whitespace
         $filename = preg_replace('/\s\s+/', ' ', $filename);
@@ -27,6 +30,10 @@ class SanitizeFilename {
         $filename = trim($filename);
         $filename = str_replace(' ', '-', $filename);
         $filename = strtolower($filename);
+
+        if(!empty($extension) && !preg_match('/\s/', $extension)) {
+            $filename .= ".$extension";
+        }
 
         return $filename;
     }