+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.0.2
+
+- bugfix SanitizeFilename
+- phpunit SanitizeFilename
+
 ## 2.0.1
 
 - added changelog